Extract marker detail editing workflows
